Updated MYNT startup

+++ start_ros.sh
@@ -36,6 +36,7 @@ then
     export START_CONTROL=0
     export START_CAMERA_SYNC=0
     export START_UWB_FUSE=0
+    export START_DJISDK=0
 
     if [ $SWARM_START_MODE -ge 1 ]
     then
@@ -63,16 +64,26 @@ then
     fi
 
 
+    if [ $START_CAMERA -eq 1 ]  && [ $CAM_TYPE -eq 0  ]  ||  [ $START_CONTROL -eq 1  ]
+    then
+        export START_DJISDK=1    
+        echo "Using Ptgrey Camera or using control, will boot dji sdk"
+    fi
+
 else
     exit 0
 fi
 
 
-if [ $START_VO_STUFF -eq 1 ]
+if [ $START_CAMERA -eq 1 ] || [ $START_VO_FUSE -eq 1] 
 then
-    echo "Enable chicken blood mode"
+    echo "Is using VO or VO FUSE, enabling chicken blood mode"
     sudo /usr/sbin/nvpmodel -m0
     sudo /home/dji/jetson_clocks.sh
+fi
+
+if [ $START_DJISDK -eq 1 ]
+then
     roslaunch dji_sdk sdk.launch &> $LOG_PATH/log_sdk.txt &
     echo "DJISDK:"$! >> $PID_FILE
     echo "Wait for DJI SDK boot up"
@@ -81,20 +92,35 @@ then
 fi
 
 
+
 if [ $START_CAMERA -eq 1 ]
 then
-    echo "Start Camera in unsync mode"
-    roslaunch swarm_vo_fuse stereo.launch is_sync:=false config_path:=$CONFIG_PATH/camera_config.yaml &> $LOG_PATH/log_camera.txt &
-    PG_PID=$!
-    echo "PTGREY_UNSYNC:"$! >> $PID_FILE
-    if [ $START_CAMERA_SYNC -eq 1 ]
+    echo "Trying to start camera driver"
+    if [ $CAM_TYPE -eq 0 ]
     then
-        sleep 5
-        sudo kill -- $PG_PID
-        echo "Start camera in sync mode"
-	    sleep 1.0
-        roslaunch swarm_vo_fuse stereo.launch is_sync:=true config_path:=$CONFIG_PATH/camera_config.yaml &>> $LOG_PATH/log_camera.txt &
-        echo "PTGREY_SYNC:"$! >> $PID_FILE
+        echo "Will use pointgrey Camera"
+        echo "Start Camera in unsync mode"
+        roslaunch swarm_vo_fuse stereo.launch is_sync:=false config_path:=$CONFIG_PATH/camera_config.yaml &> $LOG_PATH/log_camera.txt &
+        PG_PID=$!
+        echo "PTGREY_UNSYNC:"$! >> $PID_FILE
+        if [ $START_CAMERA_SYNC -eq 1 ]
+        then
+            sleep 5
+            sudo kill -- $PG_PID
+            echo "Start camera in sync mode"
+            sleep 1.0
+            roslaunch swarm_vo_fuse stereo.launch is_sync:=true config_path:=$CONFIG_PATH/camera_config.yaml &>> $LOG_PATH/log_camera.txt &
+            echo "PTGREY_SYNC:"$! >> $PID_FILE
+        fi
+    fi
+
+    if [ $CAM_TYPE -eq 1 ]
+    then
+        echo "Will use MYNT Camera"
+        source /home/dji/source/MYNT-EYE-S-SDK/wrappers/ros/devel/setup.bash
+        roslaunch mynt_eye_ros_wrapper mynteye.launch request_index:=0 &> $LOG_PATH/log_camera.txt &
+        echo "MYNT_CAMERA:"$! >> $PID_FILE
+        sleep 2
     fi
 fi
 
@@ -102,8 +128,17 @@ if [ $START_VO_STUFF -eq 1 ]
 then
     sleep 10
     echo "Image ready start VO"
-    roslaunch vins_estimator dji_stereo.launch config_path:=$CONFIG_PATH/dji_stereo/dji_stereo.yaml &> $LOG_PATH/log_vo.txt &
-    echo "VINS:"$! >> $PID_FILE
+    if [ $CAM_TYPE -eq 0 ]
+    then
+        echo "No ptgrey VINS imple yet"
+        # roslaunch vins_estimator dji_stereo.launch config_path:=$CONFIG_PATH/dji_stereo/dji_stereo.yaml &> $LOG_PATH/log_vo.txt &    
+    fi
+
+    if [ $CAM_TYPE -eq 1 ]
+    then
+        rosrun vins vins_node /home/dji/SwarmConfig/mini_mynteye_stereo/mini_mynteye_stereo_imu.yaml &> $LOG_PATH/log_vo.txt &
+        echo "VINS:"$! >> $PID_FILE
+    fi
 fi
 
 if [ $START_UWB_COMM -eq 1 ]
